small changes and cleanups
